region: Implement QueueBatch
Allows passing a batch of RPCs to the region client, which will be
added to the in progress multi and then sent to HBASE.

A note here is that it's possible for a multi to grow beyond the size
of the configured size. This feels like the right way to go because
the user of the Batch API has specifically requested the rpcs be
batched together, so splitting the batch across multiple multi's
breaks that bit of atomicity.